67. Only The Owners Can Edit or Delete Events

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -11,31 +12,23 @@ import {
   Patch,
   Post,
   Query,
+  UseGuards,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Attendee } from 'src/events/attendee.entity';
-import { Event } from 'src/events/event.entity';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
 import { EventsService } from 'src/events/events.service';
 import { CreateEventDto } from 'src/events/input/create-event.dto';
 import { ListEvents } from 'src/events/input/list.events';
 import { UpdateEventDto } from 'src/events/input/update-event.dto';
-import { Repository } from 'typeorm';
 
 @Controller('/events')
 export class EventsController {
   private readonly logger = new Logger(EventsController.name);
 
-  constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
-
-    @InjectRepository(Attendee)
-    private readonly attendeeRepository: Repository<Attendee>,
-
-    private readonly eventsService: EventsService,
-  ) {}
+  constructor(private readonly eventsService: EventsService) {}
 
   @Get()
   @UsePipes(new ValidationPipe({ transform: true }))
@@ -82,30 +75,30 @@ export class EventsController {
   //   });
   // }
 
-  @Get('practice2')
-  async practice2() {
-    // return await this.repository.findOne({
-    //   where: { id: 1 },
-    //   relations: ['attendees'],
-    // });
+  // @Get('practice2')
+  // async practice2() {
+  //   // return await this.repository.findOne({
+  //   //   where: { id: 1 },
+  //   //   relations: ['attendees'],
+  //   // });
 
-    // const event = await this.repository.findOne({ where: { id: 1 } });
+  //   // const event = await this.repository.findOne({ where: { id: 1 } });
 
-    // const attendee = new Attendee();
-    // attendee.name = 'Jerry';
-    // attendee.event = event;
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'Jerry';
+  //   // attendee.event = event;
 
-    // await this.attendeeRepository.save(attendee);
+  //   // await this.attendeeRepository.save(attendee);
 
-    // return attendee;
+  //   // return attendee;
 
-    return await this.repository
-      .createQueryBuilder('e')
-      .select(['e.id', 'e.name'])
-      .orderBy('e.id', 'ASC')
-      .take(3)
-      .getMany();
-  }
+  //   return await this.repository
+  //     .createQueryBuilder('e')
+  //     .select(['e.id', 'e.name'])
+  //     .orderBy('e.id', 'ASC')
+  //     .take(3)
+  //     .getMany();
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -120,30 +113,51 @@ export class EventsController {
   }
 
   @Post()
-  async create(@Body(ValidationPipe) input: CreateEventDto) {
-    await this.repository.save({
-      ...input,
-      when: new Date(input.when),
-    });
+  @UseGuards(AuthGuardJwt)
+  async create(@Body() input: CreateEventDto, @CurrentUser() user: User) {
+    console.log('i am here');
+
+    return await this.eventsService.createEvent(input, user);
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user,
+  ) {
+    const event = await this.eventsService.getEvent(id);
+    if (!event) {
+      throw new NotFoundException();
+    }
 
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      );
+    }
+
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id);
-    if (result?.affected !== 1) {
+  async remove(@Param('id') id, @CurrentUser() user) {
+    const event = await this.eventsService.getEvent(id);
+    if (!event) {
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to remove this event`,
+      );
+    }
+
+    return await this.eventsService.deleteEvent(id);
   }
 }
